requirements.txt
djangorestframework
django-paypal
paypalrestsdk
paypal
pillow
reportlab
matplotlib
django-crispy-forms
crispy-bootstrap4
